refactor(python): split workspace delivery owners
Goal: 2026-08-27-kungfu-python-over-1000-line-decomposition

Wave: 1/6 workspace-delivery
