+++ framework/core/src/python/kungfu/_distribution_update/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Private responsibility owners behind the distribution update facade."""
