+++ framework/core/src/python/kungfu/_workspace/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Private responsibility owners behind the public workspace facades."""
