+++ framework/core/src/python/kungfu/workspace.py
@@ -6,33 +6,60 @@ import hashlib
 import json
 import os
 import re
-import tempfile
 from dataclasses import dataclass, replace
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Literal, Mapping, cast
 from uuid import uuid4
 
 from kungfu.config import (
-    load_contract as load_config_contract,
     machine_runtime_home,
     workspace_data_home,
 )
-from kungfu.canonical_json import (
-    WORKSPACE_CANONICAL_JSON_V1,
-    canonical_json_text,
+from kungfu._workspace.continuation import (
+    _episode_bundle_root as _episode_bundle_root,
+    _full_evidence_receipts as _full_evidence_receipts,
+    inspect_workspace_continuation as inspect_workspace_continuation,
+    request_full_evidence as request_full_evidence,
+    import_full_evidence as _import_full_evidence,
+)
+from kungfu._workspace.io import (
+    _canonical_json as _canonical_json,
+    _canonical_path as _canonical_path,
+    _now as _now,
+    _semantic_root as _semantic_root,
+    _workspace_available as _workspace_available,
+    _workspace_config_home as _workspace_config_home,
+    _write_json_atomic as _write_json_atomic,
+)
+from kungfu._workspace.catalog import (
+    _CATALOG_EXCLUSION_POLICY as _CATALOG_EXCLUSION_POLICY,
+    _catalog_cut as _catalog_cut,
+    _catalog_entry as _catalog_entry,
+    _catalog_entry_key as _catalog_entry_key,
+    _catalog_lifecycle as _catalog_lifecycle,
+    _catalog_provenance as _catalog_provenance,
+    _loaded_catalog as _loaded_catalog,
+    _persisted_catalog_entry as _persisted_catalog_entry,
+    CATALOG_SCHEMA as CATALOG_SCHEMA,
+    CATALOG_ENTRY_SCHEMA as CATALOG_ENTRY_SCHEMA,
+    CATALOG_VERIFICATION_SCHEMA as CATALOG_VERIFICATION_SCHEMA,
+    CATALOG_CUT_SCHEMA as CATALOG_CUT_SCHEMA,
+    CATALOG_LIFECYCLE_PLAN_SCHEMA as CATALOG_LIFECYCLE_PLAN_SCHEMA,
+    CATALOG_LIFECYCLE_RECEIPT_SCHEMA as CATALOG_LIFECYCLE_RECEIPT_SCHEMA,
+    CatalogLifecycleState as CatalogLifecycleState,
+    load_workspace_catalog as load_workspace_catalog,
+    maintain_workspace_catalog as maintain_workspace_catalog,
+    observe_workspace_locator as observe_workspace_locator,
+    rebuild_workspace_catalog as _rebuild_workspace_catalog,
+    rebind_workspace_locator as _rebind_workspace_locator,
+    verify_workspace_catalog as _verify_workspace_catalog,
+    workspace_catalog_path as workspace_catalog_path,
 )
 
 
 WORKSPACE_SCHEMA = "kungfu.workspace.identity/v1"
 WORKSPACE_IDENTITY_MATERIAL_SCHEMA = "kungfu.workspace.identity-material/v1"
 REGISTRY_SCHEMA = "kungfu.workspace.registry/v1"
-CATALOG_SCHEMA = "kungfu.workspace.locator-catalog/v1"
-CATALOG_ENTRY_SCHEMA = "kungfu.workspace.locator-entry/v1"
-CATALOG_VERIFICATION_SCHEMA = "kungfu.workspace.locator-verification/v1"
-CATALOG_CUT_SCHEMA = "kungfu.workspace.locator-catalog-cut/v1"
-CATALOG_LIFECYCLE_PLAN_SCHEMA = "kungfu.workspace.catalog-lifecycle-plan/v1"
-CATALOG_LIFECYCLE_RECEIPT_SCHEMA = "kungfu.workspace.catalog-lifecycle-receipt/v1"
 ENSURE_RECEIPT_SCHEMA = "kungfu.workspace.ensure-receipt/v1"
 TARGET_RECEIPT_SCHEMA = "kungfu.workspace.target-receipt/v1"
 CONTINUATION_STATUS_SCHEMA = "kungfu.workspace.continuation-status/v1"
@@ -65,13 +92,6 @@ OperationClass = Literal[
     "migration",
     "destructive",
 ]
-CatalogLifecycleState = Literal["active", "retired", "test-only", "quarantined"]
-
-_CATALOG_EXCLUSION_POLICY = {
-    "retired": "explicit-retirement",
-    "test-only": "isolated-test-fixture",
-    "quarantined": "explicit-quarantine",
-}
 
 
 @dataclass(frozen=True)
@@ -88,7 +108,7 @@ class WorkspaceIdentity:
     config_home: str
 
     def as_dict(self) -> dict[str, Any]:
-        continuation = inspect_workspace_continuation(self)
+        continuation = inspect_workspace_continuation(cast(Any, self))
         return {
             "schema": WORKSPACE_SCHEMA,
             "workspace_id": self.workspace_id,
@@ -141,194 +161,10 @@ class WorkspaceTargetRequired(ValueError):
         )
 
 
-def _canonical_json(value: Any) -> str:
-    return canonical_json_text(value, protocol=WORKSPACE_CANONICAL_JSON_V1)
-
-
-def _semantic_root(value: Any) -> str:
-    digest = hashlib.sha256(_canonical_json(value).encode("utf-8")).hexdigest()
-    return f"sha256:{digest}"
-
-
 def semantic_root(value: Any) -> str:
     """Return the canonical content root used by Workspace contracts."""
 
     return _semantic_root(value)
-
-
-def inspect_workspace_continuation(identity: WorkspaceIdentity) -> dict[str, Any]:
-    """Inspect settled Git material without creating or repairing local state.
-
-    This is deliberately a bounded read model.  It can establish that tracked
-    Project Cut and Episode shadow material is present and structurally usable;
-    it never promotes that material into yijinjing authority or claims raw
-    replay/requalification evidence.
-    """
-
-    runtime_dir = Path(identity.data_home) / "runtime"
-    runtime_present = runtime_dir.is_dir()
-    episode_root = Path(identity.data_home) / "episodes" / "sealed"
-    cut_root = Path(identity.data_home) / "project-cuts"
-    issues: list[dict[str, str]] = []
-    episode_roots: list[str] = []
-    cut_roots: list[str] = []
-
-    if episode_root.is_dir():
-        for manifest_path in sorted(episode_root.rglob("manifest.json")):
-            try:
-                manifest = json.loads(manifest_path.read_text(encoding="utf-8"))
-                semantic_root = str(manifest.get("semanticRoot") or "")
-                provider_root = str(manifest.get("providerRoot") or "")
-                if (
-                    manifest.get("schema") != "kungfu.episode.git-workspace-manifest/v1"
-                    or manifest.get("authority") != "shadow-of-yijinjing-journal"
-                    or not _ROOT.fullmatch(semantic_root)
-                    or not _ROOT.fullmatch(provider_root)
-                ):
-                    raise ValueError("manifest contract mismatch")
-                provider_root_value = manifest.pop("providerRoot")
-                if _semantic_root(manifest) != provider_root_value:
-                    raise ValueError("Episode provider root mismatch")
-                episode_roots.append(semantic_root)
-            except (OSError, ValueError, json.JSONDecodeError) as error:
-                issues.append(
-                    {
-                        "code": "episode-shadow-invalid",
-                        "path": str(manifest_path.relative_to(identity.data_home)),
-                        "message": str(error),
-                    }
-                )
-
-    if cut_root.is_dir():
-        for cut_path in sorted(cut_root.rglob("*.json")):
-            if cut_path.name == "receipt.json" or cut_path.name.endswith(
-                ".receipt.json"
-            ):
-                continue
-            try:
-                cut = json.loads(cut_path.read_text(encoding="utf-8"))
-                cut_root_value = str(cut.get("cutRoot") or "")
-                if cut.get("schema") != "project.cut/v1" or not _ROOT.fullmatch(
-                    cut_root_value
-                ):
-                    raise ValueError("Project Cut contract mismatch")
-                root_input = {
-                    "schema": "project.cut.root-input/v1",
-                    **{field: cut[field] for field in _PROJECT_CUT_ROOT_FIELDS},
-                }
-                if _semantic_root(root_input) != cut_root_value:
-                    raise ValueError("Project Cut root mismatch")
-                cut_roots.append(cut_root_value)
-            except (OSError, ValueError, json.JSONDecodeError) as error:
-                issues.append(
-                    {
-                        "code": "project-cut-invalid",
-                        "path": str(cut_path.relative_to(identity.data_home)),
-                        "message": str(error),
-                    }
-                )
-
-    full_evidence_roots, full_evidence_issues = _full_evidence_receipts(
-        identity.data_home, identity.workspace_id, episode_roots
-    )
-    shadow_present = bool(episode_roots or cut_roots)
-    historical_evidence_complete = bool(episode_roots) and set(episode_roots).issubset(
-        full_evidence_roots
-    )
-    if issues:
-        state = "evidence-degraded"
-        evidence_level = "degraded"
-    elif runtime_present:
-        state = "live-runtime"
-        evidence_level = "live-local"
-    elif shadow_present:
-        state = "shadow-only"
-        evidence_level = "settled-review"
-    else:
-        state = "uninitialized"
-        evidence_level = "none"
-
-    return {
-        "schema": CONTINUATION_STATUS_SCHEMA,
-        "state": state,
-        "runtime_authority": "yijinjing-journal" if runtime_present else None,
-        "settled_history_authority": (
-            "qualified-git-shadow" if shadow_present else None
-        ),
-        "evidence_level": evidence_level,
-        "episode_roots": sorted(set(episode_roots)),
-        "project_cut_roots": sorted(set(cut_roots)),
-        "issues": issues,
-        "full_evidence_episode_roots": sorted(full_evidence_roots),
-        "full_evidence_issues": full_evidence_issues,
-        "capability_contractions": (
-            []
-            if not shadow_present or historical_evidence_complete
-            else [
-                "raw-replay-unavailable-for-settled-history",
-                "requalification-unavailable-for-settled-history",
-                "disaster-recovery-unavailable-for-settled-history",
-            ]
-        ),
-        "capabilities": {
-            "inspect_settled_history": shadow_present and not issues,
-            "start_continuation": identity.workspace_kind == "project" and not issues,
-            "append_facts": runtime_present and not issues,
-            "raw_replay": runtime_present
-            and (not episode_roots or historical_evidence_complete),
-            "requalify": runtime_present
-            and (not episode_roots or historical_evidence_complete),
-            "disaster_recovery": runtime_present
-            and (not episode_roots or historical_evidence_complete),
-            "request_full_evidence": shadow_present and not runtime_present,
-            "settle_project_cut": runtime_present and not issues,
-        },
-        "non_claims": [
-            "git-shadow-is-not-episode-authority",
-            "settled-review-does-not-prove-raw-replay",
-            "inspection-does-not-initialize-runtime",
-        ],
-    }
-
-
-def request_full_evidence(
-    identity: WorkspaceIdentity,
-    *,
-    episode_roots: list[str] | None = None,
-    project_cut_roots: list[str] | None = None,
-) -> dict[str, Any]:
-    """Create one exact, read-only request for missing local Episode evidence."""
-
-    continuation = inspect_workspace_continuation(identity)
-    if continuation["issues"]:
-        raise ValueError("full evidence cannot be requested from degraded shadow state")
-    available_episodes = set(continuation["episode_roots"])
-    available_cuts = set(continuation["project_cut_roots"])
-    if not available_episodes:
-        raise ValueError("full evidence request requires a settled Episode shadow")
-    requested_episodes = sorted(set(episode_roots or available_episodes))
-    requested_cuts = sorted(set(project_cut_roots or available_cuts))
-    if not set(requested_episodes).issubset(available_episodes):
-        raise ValueError("requested Episode root is not present in settled history")
-    if not set(requested_cuts).issubset(available_cuts):
-        raise ValueError("requested Project Cut root is not present in settled history")
-    missing = sorted(
-        set(requested_episodes) - set(continuation["full_evidence_episode_roots"])
-    )
-    plan = {
-        "schema": EVIDENCE_REQUEST_SCHEMA,
-        "workspace_id": identity.workspace_id,
-        "workspace_root": identity.workspace_root,
-        "episode_roots": requested_episodes,
-        "project_cut_roots": requested_cuts,
-        "missing_episode_roots": missing,
-        "required_bundle_schema": "kungfu.storage.episode-bundle/v1",
-        "authority": "yijinjing-journal",
-        "settled_history_authority": "qualified-git-shadow",
-        "creates_runtime": False,
-        "next_action": "obtain and validate one full Episode bundle per missing root",
-    }
-    return {**plan, "plan_root": _semantic_root(plan)}
 
 
 def import_full_evidence(
@@ -337,127 +173,13 @@ def import_full_evidence(
     *,
     execute: bool = False,
 ) -> dict[str, Any]:
-    """Validate or import a full Episode bundle bound to one settled shadow root."""
+    """Validate or import evidence through the workspace write boundary."""
 
-    from kungfu.storage import service
-
-    continuation = inspect_workspace_continuation(identity)
-    if continuation["issues"]:
-        raise ValueError("full evidence import is blocked by degraded shadow state")
-    with open(bundle_path, encoding="utf-8") as stream:
-        bundle = json.load(stream)
-    episode_root = _episode_bundle_root(bundle)
-    if episode_root not in continuation["episode_roots"]:
-        raise ValueError("Episode bundle root is not present in settled history")
-    with tempfile.TemporaryDirectory(prefix="kungfu-full-evidence-validate-") as root:
-        validation = service.import_bundle(root, bundle, verify=True, execute=False)
-    if not validation.get("ok"):
-        raise ValueError("full Episode bundle validation failed")
-    bundle_hash = (
-        "sha256:"
-        + hashlib.sha256(
-            json.dumps(
-                bundle, ensure_ascii=False, sort_keys=True, separators=(",", ":")
-            ).encode("utf-8")
-        ).hexdigest()
+    return _import_full_evidence(
+        cast(Any, identity),
+        bundle_path,
+        execute=execute,
     )
-    plan = {
-        "schema": EVIDENCE_IMPORT_PLAN_SCHEMA,
-        "workspace_id": identity.workspace_id,
-        "episode_root": episode_root,
-        "bundle_hash": bundle_hash,
-        "bundle_schema": bundle["schema"],
-        "would_create_runtime": not os.path.isdir(
-            os.path.join(identity.data_home, "runtime")
-        ),
-    }
-    plan_root = _semantic_root(plan)
-    if not execute:
-        return {
-            **plan,
-            "plan_root": plan_root,
-            "executed": False,
-            "validation": validation,
-        }
-
-    ensure_receipt = ensure_workspace_data_home(identity, "import-full-evidence")
-    imported = service.import_bundle(
-        os.path.join(identity.data_home, "runtime"),
-        bundle,
-        verify=True,
-        execute=True,
-    )
-    if not imported.get("ok"):
-        raise ValueError("full Episode bundle import failed")
-    receipt = {
-        "schema": EVIDENCE_IMPORT_RECEIPT_SCHEMA,
-        "plan_root": plan_root,
-        "workspace_id": identity.workspace_id,
-        "episode_root": episode_root,
-        "bundle_hash": bundle_hash,
-        "import_status": str(imported.get("status") or "ok"),
-        "workspace_ensure_receipt_id": ensure_receipt["receipt_id"],
-    }
-    receipt["receipt_root"] = _semantic_root(receipt)
-    receipt_path = os.path.join(
-        identity.data_home,
-        "runtime",
-        "full-evidence",
-        episode_root.removeprefix("sha256:") + ".receipt.json",
-    )
-    _write_json_atomic(receipt_path, receipt)
-    return {
-        "schema": EVIDENCE_IMPORT_RECEIPT_SCHEMA,
-        "executed": True,
-        "plan": {**plan, "plan_root": plan_root},
-        "receipt": {**receipt, "receipt_path": receipt_path},
-        "validation": validation,
-        "import": imported,
-        "continuation": inspect_workspace_continuation(identity),
-    }
-
-
-def _episode_bundle_root(bundle: dict[str, Any]) -> str:
-    if bundle.get("schema") != "kungfu.storage.episode-bundle/v1":
-        raise ValueError("full evidence must use the Episode bundle schema")
-    content_root = str((bundle.get("manifest") or {}).get("content_root") or "")
-    if re.fullmatch(r"[0-9a-f]{64}", content_root):
-        content_root = f"sha256:{content_root}"
-    if not _ROOT.fullmatch(content_root):
-        raise ValueError("Episode bundle has no valid semantic root")
-    return content_root
-
-
-def _full_evidence_receipts(
-    data_home: str, workspace_id: str, episode_roots: list[str]
-) -> tuple[set[str], list[dict[str, str]]]:
-    root = Path(data_home) / "runtime" / "full-evidence"
-    admitted: set[str] = set()
-    issues: list[dict[str, str]] = []
-    if not root.is_dir():
-        return admitted, issues
-    for path in sorted(root.glob("*.receipt.json")):
-        try:
-            receipt = json.loads(path.read_text(encoding="utf-8"))
-            receipt_root = str(receipt.pop("receipt_root", ""))
-            episode_root = str(receipt.get("episode_root") or "")
-            if (
-                receipt.get("schema") != EVIDENCE_IMPORT_RECEIPT_SCHEMA
-                or receipt.get("workspace_id") != workspace_id
-                or episode_root not in episode_roots
-                or _semantic_root(receipt) != receipt_root
-            ):
-                raise ValueError("full evidence receipt contract mismatch")
-            admitted.add(episode_root)
-        except (OSError, ValueError, json.JSONDecodeError) as error:
-            issues.append(
-                {
-                    "code": "full-evidence-invalid",
-                    "path": str(path.relative_to(data_home)),
-                    "message": str(error),
-                }
-            )
-    return admitted, issues
 
 
 def home_data_home(env: Mapping[str, str] | None = None) -> str:
@@ -476,233 +198,6 @@ def workspace_registry_path(
     )
 
 
-def workspace_catalog_path(
-    config_home: str | None = None,
-    *,
-    env: Mapping[str, str] | None = None,
-) -> str:
-    """Return the machine-local locator Catalog path.
-
-    The Catalog is deliberately separate from bounded GUI recents and from
-    every workspace's semantic authority.
-    """
-
-    return os.path.join(
-        config_home or _workspace_config_home(env),
-        "workspaces",
-        "catalog.json",
-    )
-
-
-def load_workspace_catalog(
-    config_home: str | None = None,
-    *,
-    env: Mapping[str, str] | None = None,
-) -> dict[str, Any]:
-    """Read the locator Catalog without creating or repairing it."""
-
-    path = workspace_catalog_path(config_home, env=env)
-    payload: dict[str, Any]
-    if not os.path.exists(path):
-        payload = {
-            "schema": CATALOG_SCHEMA,
-            "entries": [],
-            "epoch": 0,
-        }
-        return _loaded_catalog(payload, path)
-    try:
-        with open(path, encoding="utf-8") as stream:
-            loaded_payload = json.load(stream)
-        if (
-            not isinstance(loaded_payload, dict)
-            or loaded_payload.get("schema") != CATALOG_SCHEMA
-            or not isinstance(loaded_payload.get("entries"), list)
-        ):
-            raise ValueError("Catalog contract mismatch")
-        payload = loaded_payload
-        persisted_cut = _catalog_cut(payload)
-        entries: list[dict[str, Any]] = []
-        for entry in payload["entries"]:
-            if (
-                not isinstance(entry, dict)
-                or entry.get("schema") != CATALOG_ENTRY_SCHEMA
-            ):
-                raise ValueError("Catalog entry contract mismatch")
-            identity_state = str(entry.get("identity_state") or "qualified")
-            identity_root = str(entry.get("identity_root") or "")
-            locator_key = str(entry.get("locator_key") or "")
-            if identity_state not in {"qualified", "locator-candidate"}:
-                raise ValueError("Catalog entry identity_state is invalid")
-            if identity_state == "qualified" and not _ROOT.fullmatch(identity_root):
-                raise ValueError("Catalog entry identity_root is invalid")
-            if identity_state == "locator-candidate" and (
-                identity_root or not _ROOT.fullmatch(locator_key)
-            ):
-                raise ValueError("Catalog locator candidate is invalid")
-            lifecycle = _catalog_lifecycle(entry)
-            entries.append(
-                {
-                    **dict(entry),
-                    "provenance": _catalog_provenance(entry),
-                    "lifecycle": lifecycle,
-                    "retained": True,
-                    "required": lifecycle["state"] not in _CATALOG_EXCLUSION_POLICY,
-                    "exclusion_policy": _CATALOG_EXCLUSION_POLICY.get(
-                        lifecycle["state"]
-                    ),
-                }
-            )
-        return {
-            **payload,
-            "entries": entries,
-            "epoch": int(payload.get("epoch") or 0),
-            "issues": [],
-            "catalog_path": path,
-            "catalog_cut": persisted_cut,
-        }
-    except (OSError, ValueError, json.JSONDecodeError) as error:
-        fallback = {
-            "schema": CATALOG_SCHEMA,
-            "entries": [],
-            "epoch": 0,
-        }
-        return {
-            **fallback,
-            "issues": [
-                {
-                    "code": "catalog-invalid",
-                    "path": path,
-                    "message": str(error),
-                }
-            ],
-            "catalog_path": path,
-            "catalog_cut": _catalog_cut(fallback),
-        }
-
-
-def observe_workspace_locator(
-    identity: WorkspaceIdentity,
-    *,
-    lifecycle: CatalogLifecycleState = "active",
-    lifecycle_reason: str | None = None,
-    config_home: str | None = None,
-    env: Mapping[str, str] | None = None,
-) -> dict[str, Any]:
-    """Record one explicitly selected or successfully written locator."""
-
-    if lifecycle not in {"active", "test-only"}:
-        raise ValueError(
-            "initial Catalog observation supports only active or test-only; "
-            "use catalog-maintain for other lifecycle states"
-        )
-    lifecycle_reason = str(lifecycle_reason or "").strip()
-    if lifecycle != "active" and not lifecycle_reason:
-        raise ValueError("non-active Catalog observation requires a reason")
-
-    catalog = load_workspace_catalog(config_home, env=env)
-    if catalog["issues"]:
-        raise ValueError(
-            "invalid Workspace Locator Catalog must be repaired explicitly"
-        )
-    observed_at = _now()
-    entry = _catalog_entry(
-        identity,
-        observed_at,
-        provenance={
-            "source": "explicit-observation",
-            "registration_reason": identity.resolution_reason,
-        },
-    )
-    if lifecycle != "active":
-        entry["provenance"] = {
-            "source": "explicit-disposable-observation",
-            "registration_reason": lifecycle_reason,
-        }
-        entry["lifecycle"] = {
-            "state": lifecycle,
-            "reason": lifecycle_reason,
-            "transitioned_at": observed_at,
-            "previous_state": None,
-        }
-    existing = next(
-        (
-            row
-            for row in catalog["entries"]
-            if _catalog_entry_key(row) == _catalog_entry_key(entry)
-        ),
-        None,
-    )
-    if existing is not None:
-        if lifecycle != _catalog_lifecycle(existing)["state"]:
-            raise ValueError(
-                "existing Catalog lifecycle must change through catalog-maintain"
-            )
-        entry["lifecycle"] = _catalog_lifecycle(existing)
-        entry["provenance"] = _catalog_provenance(existing)
-        unchanged_entry = {
-            **entry,
-            "observed_at": existing.get("observed_at"),
-        }
-        conflicting_locator = any(
-            _catalog_entry_key(row) != _catalog_entry_key(entry)
-            and identity.identity_state == "qualified"
-            and row.get("locator_key") == entry["locator_key"]
-            for row in catalog["entries"]
-        )
-        if (
-            _persisted_catalog_entry(existing) == unchanged_entry
-            and not conflicting_locator
-        ):
-            # The Catalog is a locator set, not a recency log. Re-observing the
-            # same locator must not invalidate a live federation query cut.
-            payload = {
-                "schema": CATALOG_SCHEMA,
-                "entries": [
-                    _persisted_catalog_entry(row) for row in catalog["entries"]
-                ],
-                "epoch": int(catalog.get("epoch") or 0),
-            }
-            if "updated_at" in catalog:
-                payload["updated_at"] = catalog["updated_at"]
-            return {
-                **payload,
-                "catalog_path": catalog["catalog_path"],
-                "catalog_cut": catalog["catalog_cut"],
-                "observed": _persisted_catalog_entry(existing),
-                "changed": False,
-            }
-    entries: list[dict[str, Any]] = []
-    replaced = False
-    for row in catalog["entries"]:
-        if _catalog_entry_key(row) == _catalog_entry_key(entry):
-            entries.append(entry)
-            replaced = True
-            continue
-        if (
-            identity.identity_state == "qualified"
-            and row.get("locator_key") == entry["locator_key"]
-        ):
-            continue
-        entries.append(_persisted_catalog_entry(row))
-    if not replaced:
-        entries.insert(0, entry)
-    payload = {
-        "schema": CATALOG_SCHEMA,
-        "entries": entries,
-        "epoch": int(catalog.get("epoch") or 0) + 1,
-        "updated_at": observed_at,
-    }
-    path = workspace_catalog_path(config_home, env=env)
-    _write_json_atomic(path, payload)
-    return {
-        **payload,
-        "catalog_path": path,
-        "catalog_cut": _catalog_cut(payload),
-        "observed": entry,
-        "changed": True,
-    }
-
-
 def rebuild_workspace_catalog(
     workspace_roots: list[str] | None = None,
     *,
@@ -710,86 +205,12 @@ def rebuild_workspace_catalog(
     config_home: str | None = None,
     env: Mapping[str, str] | None = None,
 ) -> dict[str, Any]:
-    """Rebuild discovery only from bounded recents and explicit locators.
-
-    This never scans parent directories or the filesystem for ``.kungfu``.
-    It can repair a missing/corrupt Catalog because workspace authority remains
-    in each explicitly inspected workspace.
-    """
-
-    env = os.environ if env is None else env
-    identities: list[WorkspaceIdentity] = []
-    sources: list[dict[str, str]] = []
-    if include_recents:
-        registry = load_workspace_registry(config_home, env=env)
-        for row in registry["recent"]:
-            kind = str(row.get("workspace_kind") or "")
-            locator = row.get("workspace_root") or row.get("locator")
-            try:
-                identity = (
-                    _home_identity(env, "catalog-rebuild-recent")
-                    if kind == "home"
-                    else (
-                        _project_identity(
-                            str(locator),
-                            "catalog-rebuild-recent",
-                            env=env,
-                        )
-                        if kind == "project" and locator
-                        else None
-                    )
-                )
-            except (OSError, ValueError):
-                identity = None
-            if identity is not None:
-                identities.append(identity)
-                sources.append(
-                    {
-                        "source": "bounded-recent",
-                        "workspace_id": identity.workspace_id,
-                    }
-                )
-    for root in workspace_roots or []:
-        identity = _project_identity(root, "catalog-rebuild-explicit", env=env)
-        identities.append(identity)
-        sources.append({"source": "explicit", "workspace_id": identity.workspace_id})
-
-    observed_at = _now()
-    by_key: dict[str, dict[str, Any]] = {}
-    for identity in identities:
-        entry = _catalog_entry(
-            identity,
-            observed_at,
-            provenance={
-                "source": next(
-                    (
-                        row["source"]
-                        for row in sources
-                        if row["workspace_id"] == identity.workspace_id
-                    ),
-                    "catalog-rebuild",
-                ),
-                "registration_reason": identity.resolution_reason,
-            },
-        )
-        by_key[_catalog_entry_key(entry)] = entry
-    payload = {
-        "schema": CATALOG_SCHEMA,
-        "entries": [by_key[key] for key in sorted(by_key)],
-        "epoch": int(load_workspace_catalog(config_home, env=env).get("epoch") or 0)
-        + 1,
-        "updated_at": observed_at,
-    }
-    path = workspace_catalog_path(config_home, env=env)
-    _write_json_atomic(path, payload)
-    return {
-        **payload,
-        "catalog_path": path,
-        "catalog_cut": _catalog_cut(payload),
-        "sources": sources,
-        "filesystem_scan": False,
-        "authority": False,
-    }
+    return _rebuild_workspace_catalog(
+        workspace_roots,
+        include_recents=include_recents,
+        config_home=config_home,
+        env=env,
+    )
 
 
 def verify_workspace_catalog(
@@ -797,222 +218,10 @@ def verify_workspace_catalog(
     *,
     env: Mapping[str, str] | None = None,
 ) -> dict[str, Any]:
-    """Verify every accessible locator without changing the Catalog."""
-
-    catalog = load_workspace_catalog(config_home, env=env)
-    results: list[dict[str, Any]] = []
-    for entry in catalog["entries"]:
-        locator = entry.get("locator")
-        problem: dict[str, str] | None = None
-        try:
-            if entry["workspace_kind"] == "home":
-                identity = _home_identity(
-                    os.environ if env is None else env,
-                    "catalog-verification",
-                )
-            elif isinstance(locator, str) and locator:
-                identity = _project_identity(locator, "catalog-verification", env=env)
-            else:
-                identity = None
-        except (OSError, ValueError) as error:
-            identity = None
-            problem = {
-                "code": "workspace-identity-unreadable",
-                "message": str(error),
-            }
-        available = bool(identity and _workspace_available(identity))
-        actual_root = identity.identity_root if identity else ""
-        result = {
-            "identity_root": entry["identity_root"],
-            "workspace_id": entry["workspace_id"],
-            "available": available,
-            "identity_matches": bool(
-                available and actual_root == entry["identity_root"]
-            ),
-            "actual_identity_root": actual_root,
-        }
-        if problem is not None:
-            result["problem"] = problem
-        results.append(result)
-    ok = not catalog["issues"] and all(
-        ("problem" not in row) and (not row["available"] or row["identity_matches"])
-        for row in results
+    return _verify_workspace_catalog(
+        config_home,
+        env=env,
     )
-    return {
-        "schema": CATALOG_VERIFICATION_SCHEMA,
-        "ok": ok,
-        "catalog_path": catalog["catalog_path"],
-        "catalog_cut": catalog["catalog_cut"],
-        "epoch": catalog["epoch"],
-        "issues": catalog["issues"],
-        "entries": results,
-        "authority": False,
-        "writes": [],
-    }
-
-
-def maintain_workspace_catalog(
-    entry_keys: list[str],
-    action: Literal["retire", "test-only", "quarantine", "restore"],
-    reason: str,
-    *,
-    execute: bool = False,
-    transitioned_at: str | None = None,
-    expected_plan_root: str | None = None,
-    config_home: str | None = None,
-    env: Mapping[str, str] | None = None,
-) -> dict[str, Any]:
-    """Plan or execute explicit, reversible Catalog lifecycle transitions.
-
-    The Catalog contains locators only. This operation never touches a
-    workspace, its ``.kungfu`` authority, or any Work evidence. Dry-run is the
-    default and the execute path is bound to the exact Catalog cut it planned.
-    """
-
-    reason = reason.strip()
-    if action not in {"retire", "test-only", "quarantine", "restore"}:
-        raise ValueError("unsupported Catalog lifecycle action")
-    if not reason:
-        raise ValueError("Catalog lifecycle maintenance requires a reason")
-    if transitioned_at is not None:
-        try:
-            parsed_transition = datetime.fromisoformat(
-                transitioned_at.replace("Z", "+00:00")
-            )
-        except ValueError as error:
-            raise ValueError(
-                "Catalog lifecycle transitioned_at must be an ISO-8601 timestamp"
-            ) from error
-        if parsed_transition.tzinfo is None:
-            raise ValueError(
-                "Catalog lifecycle transitioned_at must include a timezone"
-            )
-    if expected_plan_root is not None:
-        if not execute:
-            raise ValueError(
-                "expected Catalog lifecycle plan root requires execute=True"
-            )
-        if transitioned_at is None:
-            raise ValueError(
-                "expected Catalog lifecycle plan root requires transitioned_at"
-            )
-        if not _ROOT.fullmatch(expected_plan_root):
-            raise ValueError(
-                "expected Catalog lifecycle plan root must be a SHA-256 root"
-            )
-    requested = set(entry_keys)
-    if not requested:
-        raise ValueError("Catalog lifecycle maintenance requires an entry key")
-    catalog = load_workspace_catalog(config_home, env=env)
-    if catalog["issues"]:
-        raise ValueError("invalid Workspace Locator Catalog must be repaired first")
-    by_key = {_catalog_entry_key(row): row for row in catalog["entries"]}
-    missing = sorted(requested - set(by_key))
-    if missing:
-        raise ValueError(f"Catalog entry key not found: {missing[0]}")
-
-    target_state: CatalogLifecycleState = {
-        "restore": "active",
-        "retire": "retired",
-        "test-only": "test-only",
-        "quarantine": "quarantined",
-    }[action]  # type: ignore[assignment]
-    transitioned_at = transitioned_at or _now()
-    changes: list[dict[str, Any]] = []
-    persisted_entries: list[dict[str, Any]] = []
-    for row in catalog["entries"]:
-        before = _persisted_catalog_entry(row)
-        key = _catalog_entry_key(row)
-        if key not in requested:
-            persisted_entries.append(before)
-            continue
-        after = {
-            **before,
-            "lifecycle": {
-                "state": target_state,
-                "reason": reason,
-                "transitioned_at": transitioned_at,
-                "previous_state": _catalog_lifecycle(row)["state"],
-            },
-        }
-        changes.append(
-            {
-                "entry_key": key,
-                "workspace_id": row.get("workspace_id"),
-                "identity_root": row.get("identity_root"),
-                "locator": row.get("locator"),
-                "before": before,
-                "after": after,
-            }
-        )
-        persisted_entries.append(after)
-
-    next_payload = {
-        "schema": CATALOG_SCHEMA,
-        "entries": persisted_entries,
-        "epoch": int(catalog.get("epoch") or 0) + 1,
-        "updated_at": transitioned_at,
-    }
-    plan = {
-        "schema": CATALOG_LIFECYCLE_PLAN_SCHEMA,
-        "action": action,
-        "reason": reason,
-        "catalog_path": catalog["catalog_path"],
-        "catalog_cut_before": catalog["catalog_cut"],
-        "catalog_cut_after": _catalog_cut(next_payload),
-        "epoch_before": int(catalog.get("epoch") or 0),
-        "epoch_after": next_payload["epoch"],
-        "changes": changes,
-        "workspace_writes": 0,
-        "authority_writes": 0,
-        "rollback": "restore each before entry from this receipt at the recorded cut",
-    }
-    plan_root = _semantic_root(
-        {key: value for key, value in plan.items() if key != "catalog_path"}
-    )
-    if expected_plan_root is not None and plan_root != expected_plan_root:
-        raise ValueError(
-            "Catalog lifecycle plan does not match expected dry-run plan root"
-        )
-    if not execute:
-        return {
-            **plan,
-            "plan_root": plan_root,
-            "executed": False,
-            "writes": [],
-        }
-
-    current = load_workspace_catalog(config_home, env=env)
-    if current["catalog_cut"] != plan["catalog_cut_before"]:
-        raise ValueError("Catalog changed after planning; retry from a fresh dry-run")
-    _write_json_atomic(catalog["catalog_path"], next_payload)
-    receipt_body = {
-        "schema": CATALOG_LIFECYCLE_RECEIPT_SCHEMA,
-        "plan_root": plan_root,
-        "action": action,
-        "reason": reason,
-        "catalog_cut_before": plan["catalog_cut_before"],
-        "catalog_cut_after": plan["catalog_cut_after"],
-        "epoch_before": plan["epoch_before"],
-        "epoch_after": plan["epoch_after"],
-        "changes": changes,
-        "workspace_writes": 0,
-        "authority_writes": 0,
-    }
-    receipt = {**receipt_body, "receipt_root": _semantic_root(receipt_body)}
-    receipt_path = os.path.join(
-        os.path.dirname(catalog["catalog_path"]),
-        "receipts",
-        receipt["receipt_root"].removeprefix("sha256:") + ".json",
-    )
-    _write_json_atomic(receipt_path, receipt)
-    return {
-        **plan,
-        "plan_root": plan_root,
-        "executed": True,
-        "receipt": {**receipt, "receipt_path": receipt_path},
-        "writes": [catalog["catalog_path"], receipt_path],
-    }
 
 
 def rebind_workspace_locator(
@@ -1022,15 +231,9 @@ def rebind_workspace_locator(
     config_home: str | None = None,
     env: Mapping[str, str] | None = None,
 ) -> dict[str, Any]:
-    """Rebind one exact identity to a moved locator after verification."""
-
-    if not _ROOT.fullmatch(identity_root):
-        raise ValueError("identity_root must be a SHA-256 root")
-    identity = _project_identity(workspace_root, "catalog-rebind", env=env)
-    if identity.identity_root != identity_root:
-        raise ValueError("new locator does not contain the expected workspace identity")
-    return observe_workspace_locator(
-        identity,
+    return _rebind_workspace_locator(
+        identity_root,
+        workspace_root,
         config_home=config_home,
         env=env,
     )
@@ -1231,7 +434,7 @@ def select_workspace(
 ) -> dict[str, Any]:
     registry = load_workspace_registry(config_home, env=env)
     selected = identity.as_dict()
-    selected["available"] = _workspace_available(identity)
+    selected["available"] = _workspace_available(cast(Any, identity))
     selected["selected_at"] = _now()
     recent = [
         item
@@ -1248,7 +451,7 @@ def select_workspace(
     path = workspace_registry_path(config_home, env=env)
     _write_json_atomic(path, payload)
     observe_workspace_locator(
-        identity,
+        cast(Any, identity),
         config_home=config_home,
         env=env,
     )
@@ -1323,7 +526,7 @@ def ensure_workspace_data_home(
             "initial Catalog observation supports only active or test-only; "
             "use catalog-maintain for other lifecycle states"
         )
-    continuation = inspect_workspace_continuation(identity)
+    continuation = inspect_workspace_continuation(cast(Any, identity))
     previous_state = continuation["state"]
     if previous_state == "evidence-degraded":
         raise ValueError(
@@ -1347,7 +550,7 @@ def ensure_workspace_data_home(
         created_paths.append(runtime_dir)
     try:
         catalog_observation = observe_workspace_locator(
-            qualified,
+            cast(Any, qualified),
             lifecycle=cast(CatalogLifecycleState, requested_catalog_lifecycle),
             lifecycle_reason=(
                 reason if requested_catalog_lifecycle != "active" else None
@@ -1530,174 +733,3 @@ def _qualified_identity(identity: WorkspaceIdentity) -> WorkspaceIdentity:
             config_home=identity.config_home,
         )
     return identity
-
-
-def _catalog_entry(
-    identity: WorkspaceIdentity,
-    observed_at: str,
-    *,
-    provenance: Mapping[str, str] | None = None,
-) -> dict[str, Any]:
-    locator_key = _semantic_root(
-        {
-            "schema": "kungfu.workspace.locator-key/v1",
-            "workspace_kind": identity.workspace_kind,
-            "locator": identity.workspace_root or "home",
-        }
-    )
-    return {
-        "schema": CATALOG_ENTRY_SCHEMA,
-        "workspace_id": identity.workspace_id,
-        "identity_root": identity.identity_root,
-        "identity_state": identity.identity_state,
-        "locator_key": locator_key,
-        "workspace_kind": identity.workspace_kind,
-        "locator": identity.workspace_root,
-        "data_home": identity.data_home,
-        "available": _workspace_available(identity),
-        "observed_at": observed_at,
-        "provenance": dict(
-            provenance
-            or {
-                "source": "explicit-observation",
-                "registration_reason": identity.resolution_reason,
-            }
-        ),
-        "lifecycle": {
-            "state": "active",
-            "reason": "observed",
-            "transitioned_at": observed_at,
-            "previous_state": None,
-        },
-    }
-
-
-def _catalog_entry_key(entry: Mapping[str, Any]) -> str:
-    return str(entry.get("identity_root") or entry.get("locator_key") or "")
-
-
-def _catalog_cut(payload: Mapping[str, Any]) -> str:
-    body = {
-        "schema": CATALOG_CUT_SCHEMA,
-        "catalog": {
-            "schema": payload.get("schema"),
-            "entries": list(payload.get("entries") or []),
-            "epoch": int(payload.get("epoch") or 0),
-            "updated_at": payload.get("updated_at"),
-        },
-    }
-    return _semantic_root(body)
-
-
-def _loaded_catalog(payload: Mapping[str, Any], path: str) -> dict[str, Any]:
-    return {
-        **dict(payload),
-        "entries": list(payload.get("entries") or []),
-        "epoch": int(payload.get("epoch") or 0),
-        "issues": [],
-        "catalog_path": path,
-        "catalog_cut": _catalog_cut(payload),
-    }
-
-
-def _catalog_provenance(entry: Mapping[str, Any]) -> dict[str, str]:
-    value = entry.get("provenance")
-    if isinstance(value, Mapping):
-        source = str(value.get("source") or "legacy")
-        reason = str(value.get("registration_reason") or "legacy-observation")
-    else:
-        source = "legacy"
-        reason = "legacy-observation"
-    return {"source": source, "registration_reason": reason}
-
-
-def _catalog_lifecycle(entry: Mapping[str, Any]) -> dict[str, Any]:
-    value = entry.get("lifecycle")
-    state = str((value or {}).get("state") or "active")
-    if state not in {"active", "retired", "test-only", "quarantined"}:
-        raise ValueError("Catalog entry lifecycle state is invalid")
-    return {
-        "state": state,
-        "reason": str((value or {}).get("reason") or "legacy-default-active"),
-        "transitioned_at": str(
-            (value or {}).get("transitioned_at") or entry.get("observed_at") or ""
-        ),
-        "previous_state": (value or {}).get("previous_state"),
-    }
-
-
-def _persisted_catalog_entry(entry: Mapping[str, Any]) -> dict[str, Any]:
-    return {
-        key: value
-        for key, value in entry.items()
-        if key
-        not in {
-            "retained",
-            "required",
-            "exclusion_policy",
-        }
-    }
-
-
-def _canonical_path(value: str) -> str:
-    return os.path.realpath(os.path.abspath(os.path.expanduser(value)))
-
-
-def _workspace_config_home(env: Mapping[str, str] | None = None) -> str:
-    """Resolve config Home against the supplied environment mapping.
-
-    ``os.path.expanduser`` only observes the process environment. Workspace
-    APIs deliberately accept isolated environment mappings, so a contract
-    default beginning with ``~`` must instead use that mapping's ``HOME``.
-    """
-
-    process_environment = env is None
-    env = os.environ if env is None else env
-    resolution = load_config_contract(env=env)["resolution"]
-    config_home_env = str(resolution["configHomeEnv"])
-    configured_value = env.get(config_home_env)
-    if (
-        process_environment
-        and not configured_value
-        and os.environ.get("PYTEST_CURRENT_TEST")
-    ):
-        configured_value = os.environ.get("KF_PYTEST_CONFIG_HOME") or os.path.join(
-            tempfile.gettempdir(),
-            f"kungfu-pytest-config-{os.getpid()}",
-        )
-    configured = str(configured_value or resolution["defaultConfigHome"])
-    mapped_home = env.get("HOME")
-    if mapped_home and (configured == "~" or configured.startswith("~/")):
-        configured = (
-            os.path.join(mapped_home, configured[2:])
-            if configured != "~"
-            else mapped_home
-        )
-    else:
-        configured = os.path.expanduser(configured)
-    return _canonical_path(configured)
-
-
-def _workspace_available(identity: WorkspaceIdentity) -> bool:
-    if identity.workspace_kind == "project":
-        return bool(identity.workspace_root and os.path.isdir(identity.workspace_root))
-    return os.path.isdir(os.path.dirname(identity.data_home))
-
-
-def _write_json_atomic(path: str, payload: dict[str, Any]) -> None:
-    parent = os.path.dirname(path)
-    os.makedirs(parent, exist_ok=True)
-    fd, temporary = tempfile.mkstemp(prefix=".workspaces-", suffix=".json", dir=parent)
-    try:
-        with os.fdopen(fd, "w", encoding="utf-8") as f:
-            json.dump(payload, f, indent=2, sort_keys=True)
-            f.write("\n")
-        os.replace(temporary, path)
-    except BaseException:
-        if os.path.exists(temporary):
-            os.unlink(temporary)
-        raise
-
-
-def _now() -> str:
-    return datetime.now(timezone.utc).isoformat().replace("+00:00", "Z")
